Create Main.c
With this module we simulate a crossroad with a big and a small road. The purpose of the module written using C language for the ATmega4808-4809 is to create a traffic light to regulate the movement of the cars and people that want to cross the crossroad. On the small road there is a sensor which here will be implemented with function random. When the sensor detects that there is a car  waiting, sends an order to turn on the red light for the main road and green for the small road. This road does not have a smart pedestrian light, so each individual presses a button in order for the pedestrian green light to open. When the pedestrian green light is on, then in the main (big) road red light is on and in the small road green light is on