Create Main.c
In this lab exercise we will simulate the operation of a smart of a household appliance moving in the space of an empty room. It starts from one corner of the room and its purpose is to draw the outline of the room. As it moves through space it will take values ​​from a sensor that measures the distance of the device from an obstacle in front of it (here it will be simulated with a potentiometer which will be measured by the ADC).